client/piClient1-requirements.txt
Adafruit-ADS1x15==1.0.2
Adafruit-Blinka==8.35.0
adafruit-circuitpython-ads1x15==2.2.25
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-connectionmanager==1.0.1
adafruit-circuitpython-dht==4.0.4
adafruit-circuitpython-ltr390==1.1.16
adafruit-circuitpython-register==1.9.18
adafruit-circuitpython-requests==3.2.0
adafruit-circuitpython-typing==1.10.3
Adafruit-GPIO==1.0.3
Adafruit-PlatformDetect==3.62.0
Adafruit-PureIO==1.1.11
ADS1115==0.2.1
numpy==1.26.4
paho-mqtt==2.0.0
picamera==1.13
pigpio==1.78
pyftdi==0.55.0
pyserial==3.5
pyusb==1.2.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
smbus2==0.4.3
spidev==3.6
sysv_ipc==1.1.0
typing_extensions==4.10.0
